requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aniso8601==10.0.1
    # via flask-restful
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
asgiref==3.10.0
    # via flask
blinker==1.9.0
    # via flask
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   flask
    #   uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via jwt
dependency-injector==4.48.2
    # via -r requirements.in
dnspython==2.8.0
    # via pymongo
docker==7.1.0
    # via testcontainers
fastapi==0.119.0
    # via -r requirements.in
flask[async]==3.1.2
    # via
    #   -r requirements.in
    #   flask-restful
flask-restful==0.3.10
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jwt==1.4.0
    # via -r requirements.in
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
motor==3.7.1
    # via odmantic
odm-p1d @ git+https://github.com/abys72/odm-p1d.git
    # via -r requirements.in
odmantic==1.0.2
    # via odm-p1d
pycparser==2.23
    # via cffi
pydantic==2.12.2
    # via
    #   fastapi
    #   odm-p1d
    #   odmantic
pydantic-core==2.41.4
    # via pydantic
pymongo==4.15.3
    # via
    #   motor
    #   odm-p1d
    #   odmantic
python-dotenv==1.1.1
    # via testcontainers
pytz==2025.2
    # via flask-restful
pywin32==311 ; sys_platform == "win32"
    # via
    #   -r requirements.in
    #   docker
requests==2.32.5
    # via docker
six==1.17.0
    # via flask-restful
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
testcontainers==4.13.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   testcontainers
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via
    #   docker
    #   requests
    #   testcontainers
uvicorn==0.37.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
wrapt==1.17.3
    # via testcontainers
